Refactor OCR services into modules
