+++ src/services/ocr/auth.ts
@@ -1,0 +1,17 @@
+import auth from '@react-native-firebase/auth';
+
+/**
+ * Retrieves the current Firebase authentication token if the user is signed in.
+ *
+ * @returns {Promise<string|null>} ID token string or `null` when no authenticated user is available.
+ */
+export const getAuthToken = async (): Promise<string | null> => {
+  try {
+    const user = auth().currentUser;
+    if (!user) return null;
+    return await user.getIdToken();
+  } catch (error) {
+    console.error('Error getting auth token:', error);
+    return null;
+  }
+};
