+++ src/services/ocr/records.ts
@@ -1,0 +1,199 @@
+import { API_URL } from '../api';
+import { normalizeKeysToSnakeCase } from '../../utils/normalize';
+import { getAuthToken } from './auth';
+import { ensureOnline, loggedFetch } from './network';
+import { extractCoffeeName } from './textHelpers';
+import type { ConfirmStructuredPayload, OCRHistory, StructuredCoffeeMetadata } from './types';
+
+/**
+ * Fetches OCR scan history for the authenticated user from the backend service.
+ *
+ * @param {number} [limit=10] - Maximum number of history items to retrieve; defaults to 10.
+ * @returns {Promise<OCRHistory[]>} Normalized list of OCR history entries or an empty array on failure.
+ */
+export const fetchOCRHistory = async (limit: number = 10): Promise<OCRHistory[]> => {
+  try {
+    await ensureOnline();
+    const token = await getAuthToken();
+    if (!token) return [];
+
+    const response = await loggedFetch(`${API_URL}/ocr/history?limit=${limit}`, {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+    });
+
+    if (!response.ok) {
+      console.warn('Failed to fetch OCR history');
+      return [];
+    }
+
+    const data = await response.json();
+
+    return data.map((item: any) => ({
+      id: item.id,
+      coffee_name: item.coffee_name || extractCoffeeName(item.corrected_text),
+      original_text: item.original_text,
+      corrected_text: item.corrected_text,
+      created_at: new Date(item.created_at),
+      rating: item.rating,
+      match_percentage: typeof item.match_percentage === 'number' ? item.match_percentage : null,
+      is_recommended: item.is_recommended,
+      is_purchased: item.is_purchased,
+      is_favorite: item.is_favorite,
+      brand: item.brand ?? null,
+      origin: item.origin ?? null,
+      roast_level: item.roast_level ?? null,
+      flavor_notes: item.flavor_notes ?? null,
+      processing: item.processing ?? null,
+      roast_date: item.roast_date ?? null,
+      varietals: item.varietals ?? null,
+      thumbnail_url: item.thumbnail_url ?? null,
+      structured_confidence: item.structured_confidence ?? null,
+      structured_uncertainty: item.structured_uncertainty ?? null,
+      confirmed_structured_metadata: item.confirmed_structured_metadata ?? null,
+      confirmed_structured_confidence: item.confirmed_structured_confidence ?? null,
+      confirmed_structured_raw: item.confirmed_structured_raw ?? null,
+    }));
+  } catch (error) {
+    console.error('Error fetching OCR history:', error);
+    return [];
+  }
+};
+
+/**
+ * Flags a scanned coffee as purchased and optionally attaches structured metadata for later personalization.
+ *
+ * @param {string} ocrLogId - Identifier of the OCR log entry to update.
+ * @param {string} coffeeName - User-friendly coffee name displayed in history and analytics.
+ * @param {string} [brand] - Optional roaster or brand to persist alongside the purchase event.
+ * @param {Partial<StructuredCoffeeMetadata>|null} [metadata] - Additional structured details to store with the purchase record.
+ * @returns {Promise<void>} Resolves when the backend acknowledges the purchase update.
+ * @throws {Error} When the user is unauthenticated or network requests fail.
+ */
+export const markCoffeePurchased = async (
+  ocrLogId: string,
+  coffeeName: string,
+  brand?: string,
+  metadata?: Partial<StructuredCoffeeMetadata> | null,
+): Promise<void> => {
+  await ensureOnline();
+  const token = await getAuthToken();
+  if (!token) throw new Error('Nie si prihlásený');
+
+  const body: Record<string, unknown> = {
+    ocr_log_id: ocrLogId,
+    coffee_name: coffeeName,
+  };
+  if (brand) {
+    body.brand = brand;
+  }
+  if (metadata) {
+    body.metadata = metadata;
+  }
+
+  await loggedFetch(`${API_URL}/ocr/purchase`, {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(normalizeKeysToSnakeCase(body)),
+  });
+};
+
+/**
+ * Confirms AI-generated structured metadata for a specific scan and persists the user's decision.
+ *
+ * @param {string} scanId - Unique scan identifier returned from the OCR pipeline.
+ * @param {ConfirmStructuredPayload} payload - Structured metadata and confidence indicators approved by the user.
+ * @returns {Promise<boolean>} True when the confirmation was accepted by the backend; otherwise false.
+ * @throws {Error} When the user is unauthenticated or the network request fails before receiving a response.
+ */
+export const confirmStructuredScan = async (
+  scanId: string,
+  payload: ConfirmStructuredPayload
+): Promise<boolean> => {
+  await ensureOnline();
+  const token = await getAuthToken();
+  if (!token) throw new Error('Nie si prihlásený');
+
+  const response = await loggedFetch(
+    `${API_URL}/ocr/${encodeURIComponent(scanId)}/structured/confirm`,
+    {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(normalizeKeysToSnakeCase(payload ?? {})),
+    }
+  );
+
+  return response.ok;
+};
+
+/**
+ * Deletes an OCR record belonging to the signed-in user.
+ *
+ * @param {string} id - Identifier of the OCR log to remove.
+ * @returns {Promise<boolean>} True when the backend deletion succeeds; false on error or missing auth.
+ */
+export const deleteOCRRecord = async (id: string): Promise<boolean> => {
+  try {
+    await ensureOnline();
+    const token = await getAuthToken();
+    if (!token) return false;
+
+    const response = await loggedFetch(`${API_URL}/ocr/${id}`, {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+    });
+
+    console.log('📥 [BE] Delete status:', response.status);
+    return response.ok;
+  } catch (error) {
+    console.error('Error deleting OCR record:', error);
+    return false;
+  }
+};
+
+/**
+ * Submits a user rating for a scanned coffee, storing sentiment for recommendation tuning.
+ *
+ * @param {string} scanId - Coffee identifier associated with the scan or recipe evaluation.
+ * @param {number} rating - Rating value typically on a bounded scale; expected to be between 1 and 5.
+ * @returns {Promise<boolean>} True when the rating was accepted by the API; otherwise false.
+ */
+export const rateOCRResult = async (scanId: string, rating: number): Promise<boolean> => {
+  try {
+    await ensureOnline();
+    const token = await getAuthToken();
+    if (!token) return false;
+
+    const response = await loggedFetch(`${API_URL}/coffee/rate`, {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(
+        normalizeKeysToSnakeCase({
+          coffee_id: scanId,
+          rating: rating,
+        })
+      ),
+    });
+
+    console.log('📥 [BE] Rate status:', response.status);
+    return response.ok;
+  } catch (error) {
+    console.error('Error rating coffee:', error);
+    return false;
+  }
+};
