+++ src/services/ocr/textDetection.ts
@@ -1,0 +1,74 @@
+const COFFEE_KEYWORDS = [
+  'coffee',
+  'káva',
+  'kava',
+  'kávy',
+  'kavy',
+  'cafe',
+  'espresso',
+  'cappuccino',
+  'latte',
+  'arabica',
+  'robusta',
+  'ristretto',
+  'macchiato',
+  'filter',
+  'brew',
+  'bean',
+  'beans',
+  'zrn',
+  'zrnk',
+  'mleta kava',
+  'mleta káva',
+  'zrnková káva',
+  'zrnkovu kavu',
+  'prazená káva',
+  'prazenu kavu',
+  'instant coffee',
+  'ground coffee',
+  'coffee beans',
+  'coffee bag',
+  'coffee package',
+  'coffee pack',
+  'balik kavy',
+  'balik kava',
+  'balenie kavy',
+  'balenie kávy',
+];
+
+const COFFEE_KEYWORD_COMBINATIONS: string[][] = [
+  ['balik', 'kav'],
+  ['balenie', 'kav'],
+  ['bag', 'coffee'],
+  ['package', 'coffee'],
+  ['pack', 'coffee'],
+  ['zrn', 'kav'],
+  ['prazen', 'kav'],
+];
+
+/**
+ * Determines whether provided text likely refers to coffee products by matching keywords and token combinations.
+ *
+ * @param {string|null|undefined} input - OCR-extracted or user-provided text to analyze.
+ * @returns {boolean} True when coffee-related terms are detected; otherwise false.
+ */
+export const isCoffeeRelatedText = (input?: string | null): boolean => {
+  if (!input) return false;
+  const normalized = input
+    .normalize('NFD')
+    .replace(/[\u0300-\u036f]/g, '')
+    .toLowerCase();
+
+  if (COFFEE_KEYWORDS.some(keyword => normalized.includes(keyword))) {
+    return true;
+  }
+
+  const tokens = normalized.split(/[^a-z0-9]+/).filter(Boolean);
+  if (tokens.length === 0) {
+    return false;
+  }
+
+  return COFFEE_KEYWORD_COMBINATIONS.some(combination =>
+    combination.every(part => tokens.some(token => token.includes(part))),
+  );
+};
