+++ src/services/ocr/textHelpers.ts
@@ -1,0 +1,27 @@
+/**
+ * Extracts a likely coffee name from raw OCR text by prioritizing known brands and limiting length.
+ *
+ * @param {string} text - OCR text to inspect; may contain brand names or arbitrary words.
+ * @returns {string} Shortened coffee name candidate or default label when detection fails.
+ */
+export const extractCoffeeName = (text: string): string => {
+  if (!text) return 'Neznáma káva';
+
+  // Hľadaj známe značky
+  const brands = [
+    'Lavazza', 'Illy', 'Segafredo', 'Kimbo', 'Pellini', 'Bazzara',
+    'Nespresso', 'Starbucks', 'Costa', 'Tchibo', 'Jacobs', 'Douwe Egberts'
+  ];
+
+  for (const brand of brands) {
+    const regex = new RegExp(`${brand}[\\s\\w]+`, 'i');
+    const match = text.match(regex);
+    if (match) {
+      return match[0].substring(0, 50);
+    }
+  }
+
+  // Ak nenájde značku, vráť prvé slová
+  const words = text.split(/\s+/).filter(w => w.length > 2);
+  return words.slice(0, 3).join(' ').substring(0, 50) || 'Neznáma káva';
+};
