+++ src/services/ocr/utils.ts
@@ -1,0 +1,18 @@
+/**
+ * Safely attempts to parse unknown JSON values without throwing.
+ *
+ * @template T
+ * @param {unknown} value - Raw value that may already be parsed or a JSON string.
+ * @returns {T|null} Parsed JSON cast to `T` when possible; otherwise `null` when parsing fails or input type is invalid.
+ */
+export const safeParseJSON = <T>(value: unknown): T | null => {
+  if (value == null) return null;
+  if (typeof value === 'object') return value as T;
+  if (typeof value !== 'string') return null;
+  try {
+    return JSON.parse(value) as T;
+  } catch (error) {
+    console.warn('Failed to parse JSON payload', error);
+    return null;
+  }
+};
